+++ naturalspeech2_pytorch/utils/expand/abbreviations.csv
@@ -1,0 +1,77 @@
+abbreviation,expansion,language
+mrs.,misess,en
+mr.,mister,en
+dr.,doctor,en
+st.,saint,en
+co.,company,en
+jr.,junior,en
+maj.,major,en
+gen.,general,en
+drs.,doctors,en
+rev.,reverend,en
+lt.,lieutenant,en
+hon.,honorable,en
+sgt.,sergeant,en
+capt.,captain,en
+esq.,esquire,en
+ltd.,limited,en
+col.,colonel,en
+ft.,fort,en
+dept.,department,en
+gov.,governor,en
+mgr.,manager,en
+prof.,professor,en
+res.,research,en
+asst.,assistant,en
+assoc.,associate,en
+info.,information,en
+mons.,monsieur,fr
+mme.,madame,fr
+dr.,docteur,fr
+co.,compagnie,fr
+jr.,junior,fr
+maj.,majeur,fr
+gen.,général,fr
+drs.,docteurs,fr
+rev.,révérend,fr
+lt.,lieutenant,fr
+hon.,honoré,fr
+sgt.,sergent,fr
+capt.,capitaine,fr
+esq.,écuyer,fr
+ltd.,limitée,fr
+col.,colonel,fr
+ft.,fort,fr
+dept.,département,fr
+gov.,gouverneur,fr
+mgr.,directeur,fr
+prof.,professeur,fr
+res.,recherche,fr
+asst.,assistant,fr
+assoc.,associé,fr
+info.,information,fr
+sr.,señor,es
+sra.,señora,es
+dr.,doctor,es
+st.,santo,es
+co.,compañía,es
+jr.,junior,es
+maj.,mayor,es
+gen.,general,es
+rev.,reverendo,es
+lt.,teniente,es
+hon.,honorable,es
+sgt.,sargento,es
+capt.,capitán,es
+esq.,esquire,es
+ltd.,limitada,es
+col.,coronel,es
+ft.,fortaleza,es
+dept.,departamento,es
+gov.,gobernador,es
+mgr.,gerente,es
+prof.,profesor,es
+res.,investigación,es
+asst.,asistente,es
+assoc.,asociado,es
+info.,información,es
